Enable default buffer donation for gradient accumulation (#8758)
